Move `ClusterError` and `RuntimeError` to their own modules
